fix(jobs): fixed the jobs api of scheduleAnalysis
